Refactoring and adding new update testing
Fixed lint errors

completed webui update test and extending upgrade coverage

corrected test generation for streams

fixed ansible plrepo create

Removed unused imports

added fixes to ha ad replicated fixtures

Removed unused imports

fixed ha test to support 5.7 configuration

Fixed lint issues

removed blank line

Fix to upgrade checking definition

Fix to single appliance upgrade function

Made fixes to ha failover

Added support for new cfme version ha config

Added fixes for 5.9.3.1 support

fix to ha fixtures

Moved duplicate fixtures to central location

Made requested changes

Fix to line indentations

Split SCAP hardening into multiple functions

rebase

fixing a few tests and splitting up fixtures

fix to upgrade tests

Broke down fixtures to save on code

updated fixtures

Fix to provision request

fixed broken prvosion fixture

removed typo

fix to preconfigured pick

Removed support for older version as they have been removed from sprout

Change to db partition extend to ignore streams

Fix to extend db change

Removed old versions and unnecessary code

Removed unnecessary checks

rebased PR

Added uncollect for upgrading 5.8 to 5.10 due to bug

Adding bz/uncollect to tests

Added vm delete

Fix to vm provisioning checks

Fix for python 3 compatibility

+++ cfme/utils/appliance/db.py
@@ -60,21 +60,22 @@ class ApplianceDB(AppliancePlugin):
 
     def extend_partition(self):
         """Extends the /var partition with DB while shrinking the unused /repo partition"""
-        if self.is_partition_extended:
-            return
-        with self.appliance.ssh_client as ssh:
-            result = ssh.run_command("df -h")
-            self.logger.info("File systems before extending the DB partition:\n{}"
-                             .format(result.output))
-            ssh.run_command("umount /repo")
-            ssh.run_command("lvreduce --force --size -9GB /dev/mapper/VG--CFME-lv_repo")
-            ssh.run_command("mkfs.xfs -f /dev/mapper/VG--CFME-lv_repo")
-            ssh.run_command("lvextend --resizefs --size +9GB /dev/mapper/VG--CFME-lv_var")
-            ssh.run_command("mount -a")
-            result = ssh.run_command("df -h")
-            self.logger.info("File systems after extending the DB partition:\n{}"
-                             .format(result.output))
-            ssh.run_command("touch /var/www/miq/vmdb/.db_partition_extended")
+        if self.appliance.version < '5.8.1':
+            if self.is_partition_extended:
+                return
+            with self.appliance.ssh_client as ssh:
+                result = ssh.run_command("df -h")
+                self.logger.info("File systems before extending the DB partition:\n{}"
+                                 .format(result.output))
+                ssh.run_command("umount /repo")
+                ssh.run_command("lvreduce --force --size -9GB /dev/mapper/VG--CFME-lv_repo")
+                ssh.run_command("mkfs.xfs -f /dev/mapper/VG--CFME-lv_repo")
+                ssh.run_command("lvextend --resizefs --size +9GB /dev/mapper/VG--CFME-lv_var")
+                ssh.run_command("mount -a")
+                result = ssh.run_command("df -h")
+                self.logger.info("File systems after extending the DB partition:\n{}"
+                                 .format(result.output))
+                ssh.run_command("touch /var/www/miq/vmdb/.db_partition_extended")
 
     def drop(self):
         """ Drops the vmdb_production database
